fix `testListMultipartUpload` in aws-sdk-go
Currently, the test doesn't complete nor abort the multipart upload
after tesing the list operation. It caused the `cleanup()` to fail since
the object can't not be deleted which causes the bucket deletion to
fail.
This PR adds a `CompleteMultipartUpload` call to commit the object so
the cleanup can successfully delete the object and bucket.

Change-Id: I72bd4d604c2658682fe69f17974df0055c476518